requirements.txt
apache-airflow==2.10.3
dbt-duckdb==1.8.3
duckdb==1.1.3
pandas==2.2.3
pyarrow==18.0.0
requests==2.32.3
